Support SOCK_DGRAM and SOCK_SEQPACKET unix domain sockets
This introduces a new testport parameter "socket_type", which can
be set either to "STREAM" (default), "DGRAM" or "SEQPACKET" in
order to determine the socket type of the underlying unix domain
socket.

Please note that SOCK_SEQPACKET unix domain sockets are present on
Linux but may not be present on other operating systems.

Signed-off-by: Harald Welte <laforge@gnumonks.org>