Add "getMsgLen" function similar to IPL4asp_PT
When a SOCK_STREAM socket is used for a message-based protocol, we
must make sure to properly de-segment the messages at their boundary.

In order to do so, we introduce a call-back function which the user
can provide in order to assist the UD_PT_PROVIDER in finding the
length of the messages processed.

TODO: Documentation is not updated yet

Signed-off-by: Harald Welte <laforge@gnumonks.org>